patient save functionality added

+++ Hospital-Backend/src/main/java/org/bhagwan/hospital/controller/PatientController.java
@@ -1,10 +1,13 @@
 package org.bhagwan.hospital.controller;
 
+import java.net.URI;
 import java.util.List;
 
 import org.bhagwan.hospital.entity.Patient;
 import org.bhagwan.hospital.service.PatientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +17,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
 @RequestMapping(value = "hospital")
@@ -30,8 +34,15 @@ public class PatientController {
 	}
 	
 	@PostMapping(value = "/patients")
-	public Patient createOrSavePatient(@RequestBody Patient newPatient) {
-		return patientService.save(newPatient);
+	public 	ResponseEntity<Void> createOrSavePatient(@RequestBody Patient newPatient) {
+		
+		Patient p = patientService.save(newPatient);
+		
+		//get current resource uri
+		URI uri = ServletUriComponentsBuilder.fromCurrentRequest()
+				.path("/{id}").buildAndExpand(p.getId()).toUri();
+		
+		return ResponseEntity.created(uri).build();
 	}
 	
 	@GetMapping(value="/patients/{id}")
@@ -40,14 +51,21 @@ public class PatientController {
 	}
 	
 	@PutMapping(value="/patients/{id}")
-	public Patient updatePatient(@RequestBody Patient newPatient, @PathVariable Long id) {
-		return patientService.update(newPatient);
+	public ResponseEntity<Patient> updatePatient(@RequestBody Patient newPatient, @PathVariable Long id) {
+		
+		patientService.update(newPatient);		
+		return new ResponseEntity<Patient>(newPatient,HttpStatus.OK);
 	}
 	
 	@DeleteMapping(value="/patients/{id}")
-	public Patient deletePatient(@PathVariable Long id) {
-		patientService.deleteById(id);
-		return null;
+	public ResponseEntity<Void> deletePatient(@PathVariable Long id) {
+		
+		if(patientService.findById(id)==null)
+			return ResponseEntity.notFound().build();
+		
+		patientService.deleteById(id);	
+		
+		return ResponseEntity.noContent().build();
 	}
 	
 }
